feat: refactor navigation to seperate file
